#68 - Moved the main app and menu back to their component levels, moved the remote-entry file into remote-exports
