+++ apps/spa-3ncount3r/module-federation.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   name: 'spa-3ncount3r',
   exposes: {
-    './Module': './src/remote-entry.ts',
-    './menu': './src/remote-exports/encounters-menu/encounters-menu.ts'
+    './Module': './src/remote-exports/remote-entry.ts',
+    './menu': './src/remote-exports/remote-menu.ts'
   },
 };
