--- apps/spa-3ncount3r/src/remote-entry.ts
@@ -1,1 +1,0 @@
-export { default } from './remote-exports/app/App';
