+++ apps/spa-3ncount3r/src/remote-exports/remote-entry.ts
@@ -1,0 +1,1 @@
+export { default } from '../app/App';
