+++ apps/spa-3ncount3r/src/remote-exports/remote-menu.ts
@@ -1,0 +1,1 @@
+export { default } from '../components/modules/encounters-menu/encounters-menu';
